pyproject.toml
[project]
name = "pharmagraphrag"
version = "0.1.0"
description = "GraphRAG system for querying drug interactions and adverse events using FDA data"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Jose María Ponce Bernabé", email = "jmponcebe@gmail.com" },
]
keywords = ["graphrag", "knowledge-graph", "neo4j", "pharmacology", "fda", "faers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Data processing
    "pandas>=2.2",
    "pyarrow>=18.0",
    "requests>=2.32",
    "httpx>=0.28",
    # Knowledge Graph
    "neo4j>=5.26",
    # Vector Store
    "chromadb>=0.6",
    "sentence-transformers>=3.3",
    # LLM
    "google-genai>=1.64.0",
    "ollama>=0.4",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    # NLP / Entity extraction
    "rapidfuzz>=3.14.3",
    # Utilities
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "tqdm>=4.67",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "mypy>=1.14",
    "pre-commit>=4.0",
]
ui = [
    "streamlit>=1.41",
    "streamlit-agraph>=0.0.45",
    "pyvis>=0.3",
    "plotly>=5.24",
]

[project.urls]
Repository = "https://github.com/jmponcebe/PharmaGraphRAG"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pharmagraphrag"]

# --- Ruff ---
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # Ruff-specific rules
]
ignore = ["E501", "TC001", "TC003"]

[tool.ruff.lint.isort]
known-first-party = ["pharmagraphrag"]

# --- Pytest ---
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--tb=short",
    "--cov=pharmagraphrag",
    "--cov-report=term-missing",
]

# --- MyPy ---
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"
